+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.13.5] 7th Mar, 2025
+- Added Null Checks for Web Native Campaigns Utils
+- Fixed WebInbox Runtime Issues
+- Skipped WebInbox Retries when config not present
+- Native Display KV_Pair Custom Event Paylaod Backward Compatibility
+- Enhanced session storage check for WZRK_D with fallback and error handling.
+
+
 ## [1.13.4] 7th Mar, 2025
 - Added support to preview html for custom native html
 - Fixed white background around web-popup